Relocate CoinRecord to chia_rs to match upstream
chia-blockchain deleted chia/types/coin_record.py in 2.6.0 when CoinRecord
moved into chia_rs. Move CoinRecord/CoinRecordBackwardCompatible to
chia_rs/chia-protocol/coin_record (matching the chia_rs chia-protocol crate)
and update the two importers. Type shape unchanged.
